test: gate the session race demonstration like its sibling
